SalumedX/requirements.txt
Django==5.2.7
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.7.0
python-dotenv==1.1.0
gunicorn==23.0.0
whitenoise==6.8.2
psycopg2-binary==2.9.9
dj-database-url==2.3.0